refactor(ui): extract feed/body-renderers.ts — summary/facts/narrative renderers
