+++ packages/ui/src/tabs/feed.ts
@@ -14,7 +14,6 @@ import {
 	formatTagLabel,
 	normalize,
 	parseJsonArray,
-	toTitleLabel,
 } from "../lib/format";
 import { showGlobalNotice } from "../lib/notice";
 import { setFeedScopeFilter, setFeedTypeFilter, state } from "../lib/state";
@@ -43,9 +42,8 @@ import {
 	pageNextOffset,
 	SUMMARY_PAGE_SIZE,
 } from "./feed/pagination";
-import { renderMarkdownSafe } from "./feed/sanitize";
 import { FeedSkeletonItem } from "./feed/skeleton";
-import type { FeedItem, FeedSummary, ItemViewMode } from "./feed/types";
+import type { FeedItem, ItemViewMode } from "./feed/types";
 
 /* ── Module state ─────────────────────────────────────────── */
 
@@ -206,80 +204,11 @@ import {
 
 /* ── Rendering functions ─────────────────────────────────── */
 
-function renderSummarySections(summary: FeedSummary) {
-	const preferred = [
-		"request",
-		"outcome",
-		"plan",
-		"completed",
-		"learned",
-		"investigated",
-		"next",
-		"next_steps",
-		"notes",
-	];
-	const keys = Object.keys(summary);
-	const ordered = preferred.filter((k) => keys.includes(k));
-	return ordered
-		.map((key) => {
-			const content = String(summary[key] || "").trim();
-			if (!content) return null;
-			return h(
-				"div",
-				{ className: "summary-section", key },
-				h("div", { className: "summary-section-label" }, toTitleLabel(key)),
-				h("div", {
-					className: "summary-section-content",
-					dangerouslySetInnerHTML: { __html: renderMarkdownSafe(content) },
-				}),
-			);
-		})
-		.filter(Boolean);
-}
-
-function renderFactsContent(facts: unknown[]) {
-	const trimmed = facts.map((f) => String(f || "").trim()).filter(Boolean);
-	if (!trimmed.length) return null;
-	const labeledFacts = trimmed.every((f) => /.+?:\s+.+/.test(f));
-	if (labeledFacts) {
-		const rows = trimmed
-			.map((fact, index) => {
-				const splitAt = fact.indexOf(":");
-				const labelText = fact.slice(0, splitAt).trim();
-				const contentText = fact.slice(splitAt + 1).trim();
-				if (!labelText || !contentText) return null;
-				return h(
-					"div",
-					{ className: "summary-section", key: `${labelText}-${index}` },
-					h("div", { className: "summary-section-label" }, labelText),
-					h("div", {
-						className: "summary-section-content",
-						dangerouslySetInnerHTML: { __html: renderMarkdownSafe(contentText) },
-					}),
-				);
-			})
-			.filter(Boolean);
-		if (rows.length) return h("div", { className: "feed-body facts" }, rows);
-	}
-	return h(
-		"div",
-		{ className: "feed-body" },
-		h(
-			"ul",
-			null,
-			trimmed.map((fact, index) => h("li", { key: `${fact}-${index}` }, fact)),
-		),
-	);
-}
-
-function renderNarrativeContent(narrative: string, className = "feed-body") {
-	const content = String(narrative || "").trim();
-	if (!content) return null;
-	return h("div", {
-		className,
-		dangerouslySetInnerHTML: { __html: renderMarkdownSafe(content) },
-	});
-}
+import {
+	renderFactsContent,
+	renderNarrativeContent,
+	renderSummarySections,
+} from "./feed/body-renderers";
 
 function FeedViewToggle({
 	modes,
